Fix hanging when graphing things like 1/x, x^.5, etc (technically)
This updates the line drawing method so that it will truncate the
lines it is drawing to the size of the screen. This way, it will
not hang when drawing lines from INT_MIN to INT_MAX, which is
probably what is happening when graphing 1/x, x^.5, etc - as
Inf or NaN is being converted to INT_MIN or INT_MAX or something.