Implement basic integration So I'm working on implementing integration, sums, and products, and so far I've implemented the very bare bones of integration.
Right now the syntax for integration is pretty terrible. To
integrate function f(x) from a to b (dx), type:
(a -> b) i f

So, for example, to see the integral of 2xdx from 0 to x, type:
(0 -> x) i 2x

This, of course, is not correct - really, this should be the
integral from 0 to x of a function of another variable, like 2tdt.

In the future, the integration syntax will look like this:
int (a -> b) f(t)dt